Move SpecialFieldValue to model
