+++ src/main/java/net/sf/jabref/gui/menus/RightClickMenu.java
@@ -33,7 +33,7 @@ import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
 import net.sf.jabref.preferences.JabRefPreferences;
 import net.sf.jabref.specialfields.SpecialField;
-import net.sf.jabref.specialfields.SpecialFieldValue;
+import net.sf.jabref.model.entry.SpecialFieldValue;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
