+++ src/main/java/net/sf/jabref/gui/specialfields/SpecialFieldActions.java
@@ -1,8 +1,8 @@
 package net.sf.jabref.gui.specialfields;
 
 import net.sf.jabref.gui.JabRefFrame;
+import net.sf.jabref.model.entry.SpecialFieldValue;
 import net.sf.jabref.specialfields.SpecialField;
-import net.sf.jabref.specialfields.SpecialFieldValue;
 
 public class SpecialFieldActions {
