+++ src/main/java/net/sf/jabref/gui/specialfields/SpecialFieldDropDown.java
@@ -11,13 +11,13 @@ import javax.swing.JPopupMenu;
 
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.gui.JabRefFrame;
-import net.sf.jabref.gui.specialfields.SpecialFieldIcon;
 import net.sf.jabref.logic.util.OS;
+import net.sf.jabref.model.entry.SpecialFieldValue;
+import net.sf.jabref.specialfields.SpecialField;
 
 import com.jgoodies.looks.HeaderStyle;
 import com.jgoodies.looks.Options;
-import net.sf.jabref.specialfields.SpecialField;
-import net.sf.jabref.specialfields.SpecialFieldValue;
+
 
 public class SpecialFieldDropDown {
