+++ src/main/java/net/sf/jabref/gui/specialfields/SpecialFieldIcon.java
@@ -4,9 +4,9 @@ import javax.swing.Icon;
 import javax.swing.JLabel;
 
 import net.sf.jabref.gui.IconTheme;
-
+import net.sf.jabref.model.entry.SpecialFieldValue;
 import net.sf.jabref.specialfields.SpecialField;
-import net.sf.jabref.specialfields.SpecialFieldValue;
+
 
 public class SpecialFieldIcon {
